+++ react/AffiliateURLMonitoring.tsx
@@ -6,11 +6,20 @@ import { useRuntime } from 'vtex.render-runtime'
 import SET_ON_ORDER_FORM_MUTATION from './graphql/setAffiliateOnOrderForm.graphql'
 import { DEFAULT_ORDER_FORM_ID } from './utils/constants'
 
-function AffiliateURLMonitoring() {
-  const { orderForm, setOrderForm } = useOrderForm()
-  const { route } = useRuntime()
+interface Props {
+  parameter: string
+}
 
-  const slug = route?.queryString?.targeting
+function AffiliateURLMonitoring(props: Props) {
+  const { orderForm, setOrderForm } = useOrderForm()
+  const {
+    route: { queryString },
+  } = useRuntime()
+
+  const { parameter } = props
+
+  const param = parameter || `targeting`
+  const slug = queryString != null ? queryString[param] : false
   const orderFormId = orderForm.id
 
   const [setAffiliateOnOrderForm, { called: mutationHasBeenCalled }] =
@@ -30,6 +39,20 @@ function AffiliateURLMonitoring() {
   }
 
   return <></>
+}
+
+AffiliateURLMonitoring.schema = {
+  title: 'editor.affiliateURLmonitoring.title',
+  description: 'editor.affiliateURLmonitoring.description',
+  type: 'object',
+  properties: {
+    parameter: {
+      title: 'editor.affiliateURLmonitoring.param',
+      description: 'editor.affiliateURLmonitoring.paramDetails',
+      type: 'string',
+      default: null,
+    },
+  },
 }
 
 export default AffiliateURLMonitoring